backend-data-hf/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
redis>=5.0.1
upstash-redis>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pydantic[email]>=2.5.2
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
certifi>=2024.2.2
python-multipart>=0.0.6
syncedlyrics>=0.6.0
fastapi-cache2>=0.2.1
jinja2>=3.1.2
google-genai>=0.3.0
slowapi>=0.1.9
ytmusicapi>=1.5.0
yt-dlp
curl-cffi

